requirements.txt
pymupdf
pytesseract
pillow
pandas
openpyxl
python-docx
